renames : main app escapeerrands_dj -> escapeerrands
