+++ core/tests.py
@@ -1,0 +1,2 @@
+from core.models import *
+
