+++ escapeerrands/wsgi.py
@@ -3,7 +3,7 @@ import os
 from django.core.wsgi import get_wsgi_application
 from whitenoise.django import DjangoWhiteNoise
 
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "escapeerrands_dj.settings")
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "escapeerrands.settings")
 
 application = get_wsgi_application()
 application = DjangoWhiteNoise(application)
